tf2vgg16/requirements.txt
absl-py==0.12.0
astroid==2.4.2
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
grpcio-tools==1.31.0
h5py==2.10.0
idna==2.10
isort==4.3.21
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.3
pymilvus==1.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.7.0
six==1.15.0
tensorboard==2.5.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
toml==0.10.1
typing-extensions==3.7.4.3
ujson==3.1.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
